chore: move transaction decode out
